+++ package.ps1
@@ -178,6 +178,7 @@ $firefoxManifest.background = [pscustomobject]@{
         'src/weather.js',
         'src/samsung.js',
         'src/redirect.js',
+        'src/texturl.js',
         'src/utm.js',
         'src/background.js'
     )